Add ability to add prompts
