+++ packages/react-frontend/src/App.js
@@ -8,7 +8,6 @@ import { ThemeProvider, createTheme } from "@mui/material/styles";
 import { jwtDecode } from "jwt-decode";
 
 function App() {
-  const [prompts, setPrompts] = useState([]);
   const theme = createTheme({
     palette: {
       primary: {
@@ -19,18 +18,6 @@ function App() {
       },
     },
   });
-  function fetchPrompts() {
-    const promise = fetch("http://localhost:8000/prompts");
-    return promise;
-  }
-  useEffect(() => {
-    fetchPrompts()
-      .then((res) => res.json())
-      .then((json) => setPrompts(json))
-      .catch((error) => {
-        console.log(error);
-      });
-  }, []);
 
   const [profile, setProfile] = useState({
     username: "",
@@ -42,31 +29,6 @@ function App() {
     bets: [],
   });
   const [loggedIn, setLoggedIn] = useState(false);
-
-  function postPrompt(prompt) {
-    const promise = fetch("http://localhost:8000/prompts", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(prompt),
-    });
-    return promise;
-  }
-
-  function updatePrompts(prompt) {
-    postPrompt(prompt)
-      .then((response) => {
-        if (response.status === 201) {
-          response.json().then((responseJson) => {
-            setPrompts([...prompts, responseJson]);
-          });
-        }
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  }
 
   useEffect(() => {
     const token = localStorage.getItem("token");
@@ -112,7 +74,6 @@ function App() {
             element={
               <Navigation
                 profile={profile}
-                prompts={prompts}
                 loggedIn={loggedIn}
                 setLoggedIn={setLoggedIn}
               />
