+++ packages/react-frontend/src/Navigation.js
@@ -21,22 +21,6 @@ function Navigation(props) {
     }
   }, [navigate, props.loggedIn]);
 
-  const [prompts, setPrompts] = useState([
-    { id: 1, description: "Will Cal Poly Basketball beat Hawaii?", creator: 1 },
-    { id: 2, description: "Will Cal Poly baseball beat Missouri?", creator: 2 },
-    { id: 3, description: "Will Chris Caudillo win?", creator: 3, open: true },
-  ]);
-  function updateList(prompt) {
-    setPrompts([...prompts, prompt]);
-  }
-
-  function removeOnePrompt(index) {
-    const updated = prompts.filter((prompt, i) => {
-      return i !== index;
-    });
-    setPrompts(updated);
-  }
-
   const handleSignOut = () => {
     localStorage.removeItem("token");
     props.setLoggedIn(false);
@@ -49,12 +33,8 @@ function Navigation(props) {
       <Stack padding="20px" marginBottom="60px">
         {
           [
-            <HomePage
-              profile={props.profile}
-              prompts={props.prompts}
-              removePrompt={removeOnePrompt}
-            />,
-            <CreatePage profile={props.profile} updateList={updateList} />,
+            <HomePage profile={props.profile} />,
+            <CreatePage profile={props.profile} />,
             <ProfilePage profile={props.profile} />,
           ][pageIndex]
         }
