+++ packages/react-frontend/src/pages/HomePage.js
@@ -2,8 +2,22 @@ import { Box, Typography } from "@mui/material";
 import { Prompt } from "../components/Prompt.js";
 import Feed from "../components/Feed.js";
 import Bet from "../components/Bet.js";
+import { useEffect, useState } from "react";
 
 function HomeBody(props) {
+  const [prompts, setPrompts] = useState([]);
+  function fetchPrompts() {
+    const promise = fetch("http://localhost:8000/prompts");
+    return promise;
+  }
+  useEffect(() => {
+    fetchPrompts()
+      .then((res) => res.json())
+      .then((json) => setPrompts(json))
+      .catch((error) => {
+        console.log(error);
+      });
+  }, []);
   return (
     <Box
       style={{
@@ -12,9 +26,9 @@ function HomeBody(props) {
         alignItems: "center",
       }}
     >
-      <Feed questions={props.prompts} />
+      <Feed questions={prompts} />
     </Box>
   );
-};
+}
 
 export default HomeBody;
